+++ yash-builtin/src/cd/syntax.rs
@@ -26,9 +26,9 @@ use thiserror::Error;
 use yash_env::Env;
 use yash_env::semantics::Field;
 use yash_syntax::source::Location;
-use yash_syntax::source::pretty::Annotation;
-use yash_syntax::source::pretty::AnnotationType;
-use yash_syntax::source::pretty::MessageBase;
+#[allow(deprecated)]
+use yash_syntax::source::pretty::{Annotation, AnnotationType, MessageBase};
+use yash_syntax::source::pretty::{Report, ReportType, Snippet};
 
 /// Error in parsing command line arguments
 #[derive(Clone, Debug, Eq, Error, PartialEq)]
@@ -55,6 +55,38 @@ pub enum Error {
     UnexpectedOperands(Vec<Field>),
 }
 
+impl Error {
+    /// Converts this error to a [`Report`].
+    #[must_use]
+    pub fn to_report(&self) -> Report<'_> {
+        let (location, label) = match self {
+            Self::CommonError(e) => return e.to_report(),
+            Self::EnsurePwdNotPhysical(location) => {
+                (location, "-e option must be used with -P".into())
+            }
+            Self::EmptyOperand(operand) => (&operand.origin, "empty operand".into()),
+            Self::UnexpectedOperands(operands) => (
+                &operands[0].origin,
+                format!("{}: unexpected operand", operands[0].value).into(),
+            ),
+        };
+
+        let mut report = Report::new();
+        report.r#type = ReportType::Error;
+        report.title = self.to_string().into();
+        report.snippets = Snippet::with_primary_span(location, label);
+        report
+    }
+}
+
+impl<'a> From<&'a Error> for Report<'a> {
+    #[inline]
+    fn from(error: &'a Error) -> Self {
+        error.to_report()
+    }
+}
+
+#[allow(deprecated)]
 impl MessageBase for Error {
     fn message_title(&self) -> Cow<'_, str> {
         self.to_string().into()